[v12] refactor: enable all breaking changes (#1313)
* chore: enable all breaking changes

* fix: ci issues

* chore: fix lint

* chore: cleanup

* chore: add migration guide

* chore: update docs

* chore: docs tweaks

* chore: docs tweaks

* chore: remove useBreakingChanges config flag

* refactor: code review changes

* docs: tweaks
