+++ src/queries/__tests__/text.test.tsx
@@ -503,7 +503,7 @@ test('byText support hidden option', () => {
   );
 });
 
-test('byText should return composite Text', () => {
+test('byText should return host component', () => {
   const { getByText } = render(<Text>hello</Text>);
-  expect(getByText('hello').type).toBe(Text);
+  expect(getByText('hello').type).toBe('Text');
 });
