+++ src/queries/displayValue.ts
@@ -1,9 +1,7 @@
 import type { ReactTestInstance } from 'react-test-renderer';
-import { TextInput } from 'react-native';
 import { filterNodeByType } from '../helpers/filterNodeByType';
 import { findAll } from '../helpers/findAll';
 import { matches, TextMatch, TextMatchOptions } from '../matches';
-import { getConfig } from '../config';
 import { getHostComponentNames } from '../helpers/host-component-names';
 import { makeQueries } from './makeQueries';
 import type {
@@ -26,13 +24,9 @@ const getTextInputNodeByDisplayValue = (
   const { exact, normalizer } = options;
   const nodeValue =
     node.props.value !== undefined ? node.props.value : node.props.defaultValue;
-  const shouldReturnHostTextInput = getConfig().useBreakingChanges;
-  const textInputType = shouldReturnHostTextInput
-    ? getHostComponentNames().textInput
-    : TextInput;
 
   return (
-    filterNodeByType(node, textInputType) &&
+    filterNodeByType(node, getHostComponentNames().textInput) &&
     matches(value, nodeValue, normalizer, exact)
   );
 };
