+++ src/queries/placeholderText.ts
@@ -1,8 +1,6 @@
 import type { ReactTestInstance } from 'react-test-renderer';
-import { TextInput } from 'react-native';
 import { findAll } from '../helpers/findAll';
 import { matches, TextMatch, TextMatchOptions } from '../matches';
-import { getConfig } from '../config';
 import { filterNodeByType } from '../helpers/filterNodeByType';
 import { getHostComponentNames } from '../helpers/host-component-names';
 import { makeQueries } from './makeQueries';
@@ -24,13 +22,9 @@ const getTextInputNodeByPlaceholderText = (
   options: TextMatchOptions = {}
 ) => {
   const { exact, normalizer } = options;
-  const shouldReturnHostTextInput = getConfig().useBreakingChanges;
-  const textInputType = shouldReturnHostTextInput
-    ? getHostComponentNames().textInput
-    : TextInput;
 
   return (
-    filterNodeByType(node, textInputType) &&
+    filterNodeByType(node, getHostComponentNames().textInput) &&
     matches(placeholder, node.props.placeholder, normalizer, exact)
   );
 };
