+++ src/queries/text.ts
@@ -1,10 +1,4 @@
 import type { ReactTestInstance } from 'react-test-renderer';
-import { Text } from 'react-native';
-import { getConfig } from '../config';
-import {
-  getCompositeParentOfType,
-  isHostElementForType,
-} from '../helpers/component-tree';
 import { filterNodeByType } from '../helpers/filterNodeByType';
 import { findAll } from '../helpers/findAll';
 import { getHostComponentNames } from '../helpers/host-component-names';
@@ -27,29 +21,6 @@ const queryAllByText = (
   instance: ReactTestInstance
 ): ((text: TextMatch, options?: ByTextOptions) => Array<ReactTestInstance>) =>
   function queryAllByTextFn(text, options = {}) {
-    const shouldReturnHostText = getConfig().useBreakingChanges;
-
-    // Legacy version: return composite Text
-    if (!shouldReturnHostText) {
-      const baseInstance = isHostElementForType(instance, Text)
-        ? getCompositeParentOfType(instance, Text)
-        : instance;
-
-      if (!baseInstance) {
-        return [];
-      }
-
-      const results = findAll(
-        baseInstance,
-        (node) =>
-          filterNodeByType(node, Text) && matchTextContent(node, text, options),
-        { ...options, matchDeepestOnly: true }
-      );
-
-      return results;
-    }
-
-    // vNext version: returns host Text
     return findAll(
       instance,
       (node) =>
